Added support for the member_type key & computer name memberships
Now checks groups for membership before trying to mount them
(previously would try to mount everything & silently fail on the shares
with inadequate permission).

Also added ability to specify that the membership check should use the
computer name instead of the username. This would result in the share
being mounted for any user that logs in (though it would still silently
fail for users with no authz).